LPS-155410 Rename factories
