+++ modules/apps/client-extension/client-extension-type-impl/src/main/java/com/liferay/client/extension/type/internal/factory/IFrameCETImplFactoryImpl.java
@@ -17,14 +17,16 @@ package com.liferay.client.extension.type.internal.factory;
 import com.liferay.client.extension.constants.ClientExtensionEntryConstants;
 import com.liferay.client.extension.exception.ClientExtensionEntryTypeSettingsException;
 import com.liferay.client.extension.model.ClientExtensionEntry;
-import com.liferay.client.extension.type.GlobalCSSCET;
+import com.liferay.client.extension.type.IFrameCET;
 import com.liferay.client.extension.type.factory.CETImplFactory;
-import com.liferay.client.extension.type.internal.GlobalCSSCETImpl;
+import com.liferay.client.extension.type.internal.IFrameCETImpl;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.util.UnicodeProperties;
 import com.liferay.portal.kernel.util.Validator;
 
 import java.util.Properties;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import javax.portlet.PortletRequest;
 
@@ -34,33 +36,31 @@ import org.osgi.service.component.annotations.Component;
  * @author Iván Zaera Avellón
  */
 @Component(
-	property = "type=" + ClientExtensionEntryConstants.TYPE_GLOBAL_CSS,
+	property = "type=" + ClientExtensionEntryConstants.TYPE_IFRAME,
 	service = CETImplFactory.class
 )
-public class GlobalCSSCETImplFactory implements CETImplFactory<GlobalCSSCET> {
+public class IFrameCETImplFactoryImpl implements CETImplFactory<IFrameCET> {
 
 	@Override
-	public GlobalCSSCET cet(ClientExtensionEntry clientExtensionEntry)
+	public IFrameCET cet(ClientExtensionEntry clientExtensionEntry)
 		throws PortalException {
 
-		return new GlobalCSSCETImpl(clientExtensionEntry);
+		return new IFrameCETImpl(clientExtensionEntry);
 	}
 
 	@Override
-	public GlobalCSSCET cet(PortletRequest portletRequest)
-		throws PortalException {
-
-		return new GlobalCSSCETImpl(portletRequest);
+	public IFrameCET cet(PortletRequest portletRequest) throws PortalException {
+		return new IFrameCETImpl(portletRequest);
 	}
 
 	@Override
-	public GlobalCSSCET cet(
+	public IFrameCET cet(
 			String baseURL, long companyId, String description,
 			String externalReferenceCode, String name, Properties properties,
 			String sourceCodeURL, UnicodeProperties unicodeProperties)
 		throws PortalException {
 
-		return new GlobalCSSCETImpl(
+		return new IFrameCETImpl(
 			baseURL, companyId, description, externalReferenceCode, name,
 			properties, sourceCodeURL, unicodeProperties);
 	}
@@ -71,13 +71,36 @@ public class GlobalCSSCETImplFactory implements CETImplFactory<GlobalCSSCET> {
 			UnicodeProperties oldTypeSettingsUnicodeProperties)
 		throws PortalException {
 
-		GlobalCSSCET newGlobalCSSCETImpl = new GlobalCSSCETImpl(
+		IFrameCETImpl newIFrameCETImpl = new IFrameCETImpl(
 			newTypeSettingsUnicodeProperties);
 
-		if (!Validator.isUrl(newGlobalCSSCETImpl.getURL())) {
+		Matcher matcher = _friendlyURLMappingPattern.matcher(
+			newIFrameCETImpl.getFriendlyURLMapping());
+
+		if (!matcher.matches()) {
+			throw new ClientExtensionEntryTypeSettingsException(
+				"please-enter-a-valid-friendly-url-mapping");
+		}
+
+		if (!Validator.isUrl(newIFrameCETImpl.getURL())) {
 			throw new ClientExtensionEntryTypeSettingsException(
 				"please-enter-a-valid-url");
 		}
+
+		if (oldTypeSettingsUnicodeProperties != null) {
+			IFrameCETImpl oldIFrameCETImpl = new IFrameCETImpl(
+				oldTypeSettingsUnicodeProperties);
+
+			if (newIFrameCETImpl.isInstanceable() !=
+					oldIFrameCETImpl.isInstanceable()) {
+
+				throw new ClientExtensionEntryTypeSettingsException(
+					"the-instanceable-value-cannot-be-changed");
+			}
+		}
 	}
+
+	private static final Pattern _friendlyURLMappingPattern = Pattern.compile(
+		"[A-Za-z0-9-_]*");
 
 }
